# spec/openapi.yaml
openapi: 3.0.0
info:
  title: Pico Unit API
  version: 0.4.0
  description: 'The contract served by any pico unit succesfully connected to a Wifi and running this software'
servers:
  - url: '{{baseUrl}}'
paths:
  /ping:
    get:
      summary: PING
      responses: {}
      tags:
        - monitoring
  /health:
    get:
      summary: Health
      responses:
        '200':
          description: System health snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  uptime:
                    type: object
                    properties:
                      total_s:
                        type: integer
                        description: Cumulative uptime across scheduled reboots (seconds)
                      since_boot_s:
                        type: integer
                        description: Current session uptime (seconds)
                      reboot_reason:
                        type: string
                        enum: [power_on, watchdog, scheduled, unknown]
                        description: Reason for the most recent reboot
                      scheduled_reboot:
                        type: object
                        properties:
                          enabled:
                            type: boolean
                            description: Whether the daily scheduled reboot is enabled
                          hour:
                            type: integer
                            description: Target hour for scheduled reboot (0-23, local time)
                          minute:
                            type: integer
                            description: Target minute for scheduled reboot (0-59)
      tags:
        - monitoring
  /system:
    get:
      summary: System
      responses:
        '200':
          description: System information
          content:
            application/json:
              schema:
                type: object
                properties:
                  uptime:
                    type: object
                    properties:
                      total_s:
                        type: integer
                        description: Cumulative uptime across scheduled reboots (seconds)
                      since_boot_s:
                        type: integer
                        description: Current session uptime (seconds)
                      reboot_reason:
                        type: string
                        enum: [power_on, watchdog, scheduled, unknown]
                        description: Reason for the most recent reboot
                      scheduled_reboot:
                        type: object
                        properties:
                          enabled:
                            type: boolean
                            description: Whether the daily scheduled reboot is enabled
                          hour:
                            type: integer
                            description: Target hour for scheduled reboot (0-23, local time)
                          minute:
                            type: integer
                            description: Target minute for scheduled reboot (0-59)
      tags:
        - monitoring
  /sensors:
    get:
      summary: Readings
      parameters:
        - name: force
          in: query
          required: false
          deprecated: false
          schema: {}
          example: '1'
      responses: {}
      tags:
        - sensors
  /setpoints:
    get:
      summary: Get settings
      responses: {}
      tags:
        - setpoints
    post:
      summary: Change settings
      responses: {}
      tags:
        - setpoints
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                temperature:
                  type: integer
                humidity:
                  type: integer
            example:
              temperature: ''
              humidity: ''
  /setup:
    get:
      summary: Get devices characteristics
      responses: {}
      tags:
        - setup
    post:
      summary: Change devices setup (persists to flash — survives reboot)
      responses: {}
      tags:
        - setup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pins:
                  type: object
                  properties:
                    dht:        { type: integer, minimum: 1, maximum: 29 }
                    humidifier: { type: integer, minimum: 1, maximum: 29 }
                    fan:        { type: integer, minimum: 1, maximum: 29 }
                    heater:     { type: integer, minimum: 1, maximum: 29 }
                active_high:
                  type: boolean
            example:
              pins: { dht: 4, humidifier: 6, fan: 7, heater: 8 }
              active_high: false
  /outputs:
    get:
      summary: Get outputs status
      responses: {}
      tags:
        - outputs
    post:
      summary: Change outputs status
      responses: {}
      tags:
        - outputs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fan:
                  type: boolean
                humidifier:
                  type: boolean
                heater:
                  type: boolean
            example:
              fan: true
              humidifier: false
              heater: false
  /control:
    get:
      summary: Get control loop status
      responses: {}
      tags:
        - control
    post:
      summary: Toggle control loop
      responses: {}
      tags:
        - control
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
            example:
              enabled: false
  /:
    get:
      summary: Get all the system's current parameters
      parameters:
        - name: force
          in: query
          required: false
          deprecated: false
          schema: {}
          example: '1'
      responses:
        '200':
          description: Full system snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  uptime:
                    type: object
                    properties:
                      total_s:
                        type: integer
                        description: Cumulative uptime across scheduled reboots (seconds)
                      since_boot_s:
                        type: integer
                        description: Current session uptime (seconds)
                      reboot_reason:
                        type: string
                        enum: [power_on, watchdog, scheduled, unknown]
                        description: Reason for the most recent reboot
                      scheduled_reboot:
                        type: object
                        properties:
                          enabled:
                            type: boolean
                            description: Whether the daily scheduled reboot is enabled
                          hour:
                            type: integer
                            description: Target hour for scheduled reboot (0-23, local time)
                          minute:
                            type: integer
                            description: Target minute for scheduled reboot (0-59)
  /reboot:
    post:
      summary: Trigger a reboot
      description: >-
        Initiates a soft or hard reboot. The response is returned before the
        reboot executes (the Pico becomes unreachable shortly after).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [soft, hard]
                  default: soft
                  description: >-
                    soft — machine.soft_reset() (preserves reset_cause classification).
                    hard — machine.reset() (full power-cycle, WDT-style).
            example:
              type: soft
      responses:
        '200':
          description: Reboot initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                    enum: [soft, hard]
              example:
                message: Reboot initiated
                type: soft
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                bad_json:
                  summary: Invalid JSON body
                  value:
                    error: invalid JSON body
                bad_type:
                  summary: Invalid reboot type
                  value:
                    error: "type must be 'soft' or 'hard'"
      tags:
        - system